requirements.txt
crewai>=0.55.2
crewai[tools]>=0.55.2
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
mcp>=1.0.0
notion-client>=2.2.1
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
